fix: 外部数据源配置用户冲突补充校验规则
