+++ src/pages/src/hooks/index.ts
@@ -8,3 +8,4 @@ export * from './useTableMaxHeight';
 export * from './useAdminPassword';
 export * from './useInfoBoxContent';
 export * from './useCountDown';
+export * from './useConflictRules';
